# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Documentation",
        "description": "Implementation of Swagger with in Laravel",
        "contact": {
            "email": "hamzaouhamou@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "3.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/me": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get Info Of Current User",
                "description": "Returns user's info",
                "operationId": "me",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/users": {
            "put": {
                "tags": [
                    "User"
                ],
                "summary": "Update user",
                "operationId": "fe5199ce950b39e4f6437d0eac7aafd3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "User id",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-sources": {
            "get": {
                "tags": [
                    "Source"
                ],
                "summary": "list all user-source records",
                "operationId": "2dcde1e1252c3dda946f1348cfcee713",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Source"
                ],
                "summary": "Add user-source",
                "operationId": "2c74a580ee1dbb2b0a8fe13fb79b9169",
                "requestBody": {
                    "description": "user-source object",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserSourceRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/user-sources/{id}": {
            "delete": {
                "tags": [
                    "Source"
                ],
                "summary": "delete user-source record",
                "operationId": "26d4c47a8c7fff2955abd34147c8f006",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "user-sources id",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "user-sources object",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Validation errors"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UserSourceRequest": {
                "title": "User Source Request",
                "required": [
                    "user_id",
                    "source_name",
                    "source_id"
                ],
                "properties": {
                    "user_id": {
                        "description": "Class UserSource",
                        "type": "integer"
                    },
                    "source_name": {
                        "type": "string"
                    },
                    "source_id": {
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    }
}